Fixed GetAttackRiding() to take into account the weight of the man since we have alternative races with different weights.  Use the lightest man in the unit.
Did a first cut at 'attributes' which tries to abstract out some of the
things which modify stealth, observation, and tactics as well as
entertainment/phantasmal entertainment.  Not done and COMPLETELY
untested.

Realized that getting rid of hardcoded skills is going to be even worse
than I thought.. I completely forgot about all of the goddamned 'Practice'
calls throught the code :/ I think I'm just going to give up now.  It's
getting to be past the point where this is worth the headache and all the
fnording options in the code, 90% of which are unused (and probably untested)
are making it unmaintainable.