fixed compiler warning about condition evaluating address of variable
compiler warning:
- [-Waddress] / gcc

arcadia/extra.cpp/Game::CheckVictory
`if ( faction_present )` is always true, because `faction_present` is
an int[] and degrades to an int*, which can't be a nullptr, because
it's a local variable.
Intended is `if (faction_presend[i])` (like one line above).

Fixed.

arcadia/extra.cpp:163:31: error: the address of ‘faction_present’ will always evaluate as ‘true’ [-Werror=address]