fixed compiler warning about comparing a pointer to a string literal
compiler warning:
- [-Waddress]

arcadia/economy.cpp/ARegion::TownHabitat
Condition does intend to check, if object is a bank by comparing
ObjectType.name to "Bank". This fails, because it actually tries to compare
the pointer of name to (address of) string literal of "Bank" instead of
comparing the chars (pointed to by name) to the chars of the string literal.
Comparing chars is unnecessary, because we know the actual object-type and
can check that one.

Fixed.

arcadia/economy.cpp:1238:36: error: comparison with string literal results in unspecified behavior [-Werror=address]