fixed compiler warning about uninitialized variable
compiler warning:
- [-Wsometimes-uninitialized] / clang 3.8.1 -Wall -Werror

arcadia/runorders.cpp/Game::DoSendOrder
Variable `reg` is declared, might be set or not, and
then is checked for NULL `if (!reg)`.
Without initialisation the value of `reg` is unknown.

Fixed: initialized with NULL.

arcadia/runorders.cpp:3328:19: error: variable 'reg' is used uninitialized whenever 'if' condition is false
      [-Werror,-Wsometimes-uninitialized]