study to level bug (#72)
* Code assigns a value to dir when it should be comparing.

What this piece of code is suposed to do is making less likely for land
to grow to the north when selected region is at the top of the map
(directions NW, N and NE have 75% of being ignored) and same with south
when region is at the bottom (directions SW, S and SE), but the
comparation to NDIRS-1 is wrong.

This bug exists for long ago and it is also in all other games as
arcadia, kingdoms, havilah, etc

* When a unit with an AMTS has the order "study obse 5" it stops studying
as soon as it reaches OBSE-3. This is caused because void
Game::Do1StudyOrder() checks if the target level has been reached using
Unit::GetSkill(), but this function adds modifiers to the skill.