Explain in rules that fractional food is lost.
If (and only if) amount of maintenance provided by one unit of food is
greater than the amount of maintenance required to support one peasant,
adds "(or fraction thereof)" and "Fractional portions of food may not be
shared with other units" to neworigins.html.

Also fixes a bug that put an additional comma before the list of food
items.

This fix was tested as follows:

$ make neworigins-rules
$ grep -B1 -A9 "This fee is generally" neworigins/html/neworigins.html
<This showed the expected new text.  Good.>
<The extra comma was removed.  Good.>

$ sed -i /UPKEEP_FOOD_VALUE$/s/30/10/ neworigins/rules.cpp
$ make neworigins
$ make neworigins-rules
$ grep -B1 -A9 "This fee is generally" neworigins/html/neworigins.html
<This showed the older, shorter text.  Good.>
<The extra comma was removed.  Good.>