Move tests to snapshot-tests and enhance them
Snapshot tests kind of suck in general and are more fragile, but
they are all we really have right now for testing before the unit
test changes and they can still serve some purpose for doing
end-to-end testing.

This moves the directory to be better named, updates the github
workflow to use the new directory/driver script and enhances that
script to compare the output to 'expected' output each step of
the way.