Fix Landmark ordering for events
The intent of the code was that weighter landmarks (settlments) would
take preference over fortifications, etc.   This test was backwards.
However, it was also possible that if there were two equally close
and equally weighty landmarks, the order would fluctate based solely on
how the in-memory layout of the unordered_map being used to construct
the data got laid out.   This changes this to be a stable ordering based
on distance (shorter distance), weight (higher weight), x coordinate
(westernmost preferred) and finally y coordinate (northernmost
preferred).   Exposed the comparison function to allow unit tests and
added unit tests.