Write all files (except rules)  using ostream (#148)
This changes the code to use generic ostreams to write reports
as well as game output files.  The wrap/indent/logic which used
to be part of Areport is now written as a streambuf interceptor
which inserts itself into the streams that want it.   This will
allow us to (as we go farther) to use stringstreams for testing
so that we can compare/check output in test suites.

This does not *yet* modify the rules generation, nor does it
*yet* support json output for reports (and templates) but that
will be coming soon.

The snapshot tests were modified (before my code changes) to
generate one turn with map output to make sure my changes did not
break those formats either.