Update README and genrules (#156)
This updates the readme to no longer point to travis-ci for the build.
It also adds a badge for the license. The language has been cleaned
up and the copyright brought current as well as updating the maintainers
and some other cleanup.  It also removes the travis-ci config file
.travis.yml which is no longer used.

Similar changes have been made to the generated rules to bring the
credits section of them up to date.

Since this changes the output rules files, the snapshots have been
updated with the new changes as well so that they will pass.