Sanitize existing names/descriptions and prevent new issues (#159)
Since this is a diff against the old (pre-unit test) code, I cannot
prove that this works, but I will be applying an analogous patch +
unit tests against main here shortly.   This feels correct in that
it will

a) strip out the 'real' unit/faction/object number from the name.
b) rebuild the name with the new set of legal characters (and reattach
the unit/faction/object number)
c) sanitize the descriptions to only the new set of legal characters on
readin.
d) prevent (via the change to islegal()) any new names or descriptions
from containing the bogus characters.

The non-allowed characters are now the set ()[]; instead of just
() as before.