Write the text reports from json
Rather than writing the text reports and the json reports from the
internal structures, this changes the code to build the json reports
once, and then use that to output both the json report (just dump
the json) and the text reports (utilize the json data to output the
existing textual report format).

This does change one small thing which is that now the templates
are written within the WriteReports function rather than via a
second pass.  Since the Json already contains the unit order info,
this was the cleanest way to do it.   This requires updating the
snapshots of the turn run output since the lines about writing
templates are now gone.