Fix reported bug around PHBE and illusionary dragons
Looking into this also revealed that liches suffered the same
problem/bug, so they have been fixed as well.   Additionally the
spell descriptions were incorrect in that they implied you could only
summon one dragon or lich per turn illusorily, but the code would
actually allow you to summon up to the (incorrect) limit.  Allowing this
to take a number mimics the other illusory spells, and the max a unit
can hold of the illusory objects mimics the max they can hold of the
real objects.

Additional to this, while reviewing all the summon code, I discovered
that the spell description for Bird Lore 3 was incorrectly using the
.mOut value for skeletons rather than eagles.   The actual execution
code was using the correct .mOut value, so the correct fix was to make
the spell description correct to match reality.

This does cause the snapshot turns to need to be updated since those
reflected the wrong value in the spell description.