Remove the freezing code & add city counting for winner
A common New Origins win condition is percent of cities voting.
Add a global define to enable this, and put the counting and automatic
victory declaration, as well as a times message once someone has made
good progress towards it.   While this won't be used for NO7, I suspect
it will be good to have for the future.

Also, the freezing code was highly NO6 specific, so it shouldn't be
on the master branch.