Fix iterations on static arrays
A lot of the code iterated over fixed static arrays in classes by
assuming the type of the element.  If the element changed (as likely
happened at some point from int to a char * for ManType.skills) we
could end up doing bad things.  The one notably problematic case
(caught by @ennorehling in Issue #21) was fortunately under a global
which is rarely enabled so it never caused crashed.  I changed/audited
all instances of sizeof(somearray)/sizeof(...) to make sure it was
always dividing by the sizeof the first element of the arrary.  This
guarantees correctness even if a type changes to a promotable type.
(non-promotable types would be caught by the compiler)