Some various bug fixes from old issues (#193)
* Fix a memory leak in Do1Assinate and Do1Steal

* Identified by @ennorehling in Issue #22

* Fix iterations on static arrays

A lot of the code iterated over fixed static arrays in classes by
assuming the type of the element.  If the element changed (as likely
happened at some point from int to a char * for ManType.skills) we
could end up doing bad things.  The one notably problematic case
(caught by @ennorehling in Issue #21) was fortunately under a global
which is rarely enabled so it never caused crashed.  I changed/audited
all instances of sizeof(somearray)/sizeof(...) to make sure it was
always dividing by the sizeof the first element of the arrary.  This
guarantees correctness even if a type changes to a promotable type.
(non-promotable types would be caught by the compiler)

* Fix a number of bad arraybounds checks

Many of the functions in modify.cpp had checks which were just wrong
and could lead to writes to random memory/crashes.
Modified all functions to compare to the >= actual size of array being
indexed.
Found by @ennorehlingin Issues #18, #19, #20

* Fix a bogus comparison

While the function in question (to add a town/habitat) was only called
during setup, the code was wrong.  It was comparing ItemDefs.flags to a
value from the ItemDefs.type field.  Fixed.

Found by @danextpope in Issue #10