Fix Form alias parsing misbehavior
In the current game, when checking the orders, it merely checks that
the alias *starts* with a number, but forms like the following

FORM 1a
FORM 1b

etc

would be allowed and would not error during the checker but would
during the run.   This makes them error during the checker phase as
well.