Skills/Skilllist no longer use AList
In the quest to get rid of alist and alistelem, we are now removing
it from use by the unit/faction skill lists.  We kept the skilllist
type because it had some nice utility functions, but reimagined it
as a class which privately manages a std::list and implements a
normal iterator.