Clean up usage of namepace std
It's pretty bad form to have header files do 'using namespace'
directives and pull in a lot of definitions that way.  While it
isn't too horrible in a private project it's just bad practice.

This removes all the `using namespace std;` from the header files
and declares the objects with explicit namespaces.   For the .cpp
files, if there were only a few uses of the namespace, just do the
in-place explicit naming.  If the cpp file used it in a lot of
places, those are a much more common (and expected) place to add
those sort of `using` directives.

Also removed a couple of long-commented out functions.