Update entity weight
The intent was that the entities are not transportable via magic, via
transport or via the underworld, so that they cannot be moved around
except via overland/ships. However, I had mistakenly set their weight
such that high level portal lore and teleport could be used to move
them. The easiest fix was to up their weight so that this becomes
impossible, though the side effect is they now require more ships to
transport. This shouldn't be a problem however in the scheme of things
as they larger ships can handle them in a reasonable number and even
the smaller ships can do it, with enough of them in the fleet.