Remove AList usage from all orders
This modifies all orders to no longer be ALists.  Also moves
them to be shared_ptr rather than using raw pointers with
new/delete.

In most cases, I chose to use std::vector for the list of orders
but in some (two) cases I chose std::list since there was a need
to insert at the head.  Choosing vector over list is preferred in
general as it has better speed, but it makes inserting at the head
harder.