Fix bug with assassinating MANPRODUCE units
If a unit had only MANPRODUCE items (cata/sted) it would cause a crash
when the game was run if someone attempted to assassinate that unit.

This fixes the code so that MANPRODUCE units are considers for
assassination (same as they are soldiers in normal battles).  If a
unit has men AND cata/sted, assassination *will* prioritize a man over
the inanimate object.