Fix bug with assassinating STED/CATA
If a unit contained only STED/CATA it would crash the game if
someone attempted to assassinate it.

While it's possibly non-ideal to allow assassinating inanimate
things, the way the code is currently structured it's a valid
target unit.

This change makes it work, AND will prefer to assassinate a man
over an automaton in a mixed unit.