Remove another use of AList
This makes Object no longer a child of AListElem and thus
no longer using AList.  It also removes this from a couple
of other classes (Location and Farsight) which were trivial.

As part of this, discovered some bugs in handling of other AList
removals which would have crashed in the future (unit moves
while iterating the lists was a big one).