Update NO7 victory conditions to avoid delay of game tactics
There was some discussion about factions deliberately trying
to stall the game into a never-ending state.  Changes have been
made which should prevent that.

This also adds a facility that will hopefully get used more in
the future to allow ruleset specific data to affect functions
in the main game without needing to create duplicate overridden
code.  See the rulesetSpecificData member of game, and the use
of it in neworigins/extra.cpp and in the runorders.cpp for
RunAnnihilationOrders