Remove i_rand.cpp/h
The code now uses the standard library rand mt19937 random generator. In
order to keep the same basic interface and to make sure we had a
consistant seeding, we now have an rng class which exposes static
methods and utilizes an underlying singleton object to preserve state
and have just one generator which is seeded as expected.

The functions changed slightly, with seed_random having an overload with
no seed that does (effectively) what the old seedrandomrandom used to
do. In order to seed an mt19937 generator effectively you want to
provide a lot of seed data, not just a single seed, so what I did was:

a. If you pass a seed, it creates a min_rand generator, and uses that to
create a seed_seq which keeps it consistant but provides enough data to
thoroughly seed the mt19937.
b. If you don't pass a seed, it uses std::random_device (which on
machines that support it will be truly random data, but otherwise use an
underlying prng seeded appropriately) in order to provide that seed data
for the underlying mt19937.

Once seeded, the mt19937 is used for all future random.   All call
locations of getrandom(), seedrandom(), makeRoll() etc have been changed
to call rng::get_random, rng::seed_random, and rng::make_roll
respectively.

get_random also stores the distribution object statically so that if you
request multiple random numbers in the same range it will utilize the
same distribution object which is actually the best practice rather than
relying directly on the rng and using modulo (%) which is what i_rand
did.