Update parser to no longer use AString
This rewrites a lot of the orders parsing and players.in file
parsing code to remove the need for AString.  It does create
a string_parser and a token class which behave similarly to how
AString did, but
- without always allocating/deallocating memory
- using stl types and modern c++ methods to implement.

I also extended the orders checking so that it will check spell
syntax, which had been a TODO in the code for a while that I noticed
when I started working on it.

Update test helper to handle studying skills more easily

Add .cursor, .copilot and .vscode files to the archive