Clean up some easy AString uses
Change some of the simpler functions which return AString to
return std::string instead.

More cleanup of some files formatting and removal of the old
file heading blob.

Clean up a couple of enum types as well.

Move AWrite and ADot to a logger namespace and rename and allow
overriding of the stream that gets logged to (defaults to
std::cout).  This in turn simplified some of the unit test stuff
(no longer needing to replace the std::cout output buffer).

Wrote a small bit of wrapper to allow scoped enums (enum class ...)
to work for comparisons.  This lets me make the enums more tightly
defined over time while still allowing the code to be readable
without static_casts all over the place.