Remove some unneeded std::string calls
There were a couple places where strings that didn't need
to be wrapped in std::string were being wrapped.  This
removes those and also removes all cases of
'using namespace std;'

That latter had been on my plate, and Enno doing a bunch
of it as part of the string cleanup prompted me to do
the rest of it.